fix: Fix parser to build new context for dict builtin function...allowing deeply nested context passage between partials

+++ extension/src/compiler/expressionParser.test.ts
@@ -452,7 +452,7 @@ function testComplexExpressions(vars: Map<string, TemplateVar>): [number, number
 function testEdgeCases(vars: Map<string, TemplateVar>): [number, number] {
   const tests: TestCase[] = [
     { name: 'String literal', expr: '"hello world"', expectedType: 'string' },
-    { name: 'Number literal', expr: '42', expectedType: 'float64' },
+    { name: 'Number literal', expr: '42', expectedType: 'int' },
     { name: 'Parenthesized expression', expr: '(gt .Count 5)', expectedType: 'bool' },
   ];
